+++ lib/src/widgets/mp_images_widget.dart
@@ -19,7 +19,7 @@ class MPImagesWidget extends StatelessWidget {
 
         final List<Widget> widgets = [];
         final Iterable<THXTherionImageInsertConfig> images =
-            th2FileEditController.thFile.getXTherionImageInsertConfigs();
+            th2FileEditController.thFile.getImages();
 
         for (final THXTherionImageInsertConfig image in images) {
           if (!image.isVisible) continue;